Centralize creation of test users
